requirements.txt
asgiref==3.7.2
chardet==5.1.0
django==3.2.25
django-cors-headers==4.1.0
djangorestframework==3.14.0
doxypypy==0.8.8.7
numpy==1.21.6
opencv-python==4.5.3.56
Pillow==9.5.0
psycopg2-binary==2.8.6
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
djangorestframework_simplejwt==5.1.0
celery==5.2.4
redis==5.0.3
PyYAML==6.0.1

#yolov7 packages 
matplotlib>=3.2.2
opencv-python>=4.1.1
requests>=2.23.0
scipy>=1.4.1
torch
torchvision
tqdm>=4.41.0
protobuf<4.21.3

tensorboard>=2.4.1

pandas>=1.1.4
seaborn>=0.11.0

ipython 
psutil  
thop
